separation of getServerSideProps and page components
