+++ src/AwesomeApp.UI/src/common/libs/ActionResultFactories.ts
@@ -1,4 +1,4 @@
-import { ActionResult } from "@/shared/types";
+import { ActionResult } from "@/common/types";
 
 export function createSucessfulActionResult<T>(payload: T): ActionResult<T> {
   return {
