+++ src/AwesomeApp.UI/src/common/libs/EnvironmentHelpers.ts
