+++ src/AwesomeApp.UI/src/common/types/ActionResult.ts
