+++ src/AwesomeApp.UI/src/common/types/AppSettings.ts
