+++ src/AwesomeApp.UI/src/common/types/NextAuth.ts
