+++ src/AwesomeApp.UI/src/common/types/account/AuthenticateAccount.ts
