+++ src/AwesomeApp.UI/src/pages/account/_serverProps.ts
@@ -1,0 +1,3 @@
+import { ensureAuthorized } from "@/frontend/libs";
+
+export const getServerSideProps = ensureAuthorized((_, session) => ({ props: { user: session.user} }))
