+++ src/AwesomeApp.UI/src/pages/account/index.ts
@@ -1,0 +1,2 @@
+export * from "./_serverProps"
+export * from "./_account"
