+++ src/AwesomeApp.UI/src/pages/api/account/create.ts
@@ -1,6 +1,6 @@
 import { postCreateAccount } from '@/backend/handlers/AccountHandlers'
 import { withEndpoints, withErrorHandling } from '@/backend/libs'
-import { HttpMethod } from '@/shared/types/HttpMethod'
+import { HttpMethod } from '@/common/types/HttpMethod'
 
 export default 
 withErrorHandling(
