+++ src/AwesomeApp.UI/src/pages/api/account/current.ts
@@ -1,6 +1,6 @@
 import { getCurrentAccount } from '@/backend/handlers/AccountHandlers'
 import { withAuthentication, withEndpoints, withErrorHandling } from '@/backend/libs'
-import { HttpMethod } from '@/shared/types/HttpMethod'
+import { HttpMethod } from '@/common/types/HttpMethod'
 
 export default 
 withErrorHandling(
