+++ src/AwesomeApp.UI/src/pages/create-account/index.ts
@@ -1,0 +1,1 @@
+export * from "./_createAccount"
