+++ src/AwesomeApp.UI/src/pages/login/_serverProps.ts
@@ -1,0 +1,19 @@
+import { GetServerSidePropsContext, GetServerSidePropsResult } from "next";
+import { getSession } from "next-auth/react";
+
+export async function getServerSideProps(context: GetServerSidePropsContext): Promise<GetServerSidePropsResult<{ }>> {
+    const session = await getSession(context)
+  
+    if (session) {
+      return {
+        redirect: {
+          destination: '/account',
+          permanent: false,
+        }
+      }
+    }
+  
+    return {
+      props: { }
+    }
+  }
