+++ src/AwesomeApp.UI/src/pages/login/index.ts
@@ -1,0 +1,2 @@
+export * from "./_serverProps"
+export * from "./_login"
