--- src/AwesomeApp.UI/src/shared/types/index.ts
@@ -1,9 +1,0 @@
-export * from './account/Account'
-export * from './account/AccountRole'
-export * from './account/AuthenticateAccount'
-export * from './account/CreateAccount'
-export * from './account/CustomerType'
-export * from './ActionResult'
-export * from './AppSettings'
-export * from './HttpMethod'
-export * from './NextAuth'
